cleanup: put azure buid support into its own module, reenable feature builds, use a different github connection, remove unused code
