+++ azure-pipelines-template.yml
@@ -6,16 +6,16 @@ jobs:
         toolchain: stable
         features: ''
         exampleArgs: '--driver cpu --driver pdf'
-      #beta:
-      #  toolchain: beta
-      #  features: ''
-      #stable-vulkan:
-      #  toolchain: stable
-      #  features: 'vulkan'
-      #stable-svg:
-      #  toolchain: stable
-      #  features: 'svg'
-      #  exampleArgs: '--driver svg'
+      beta:
+        toolchain: beta
+        features: ''
+      stable-vulkan:
+        toolchain: stable
+        features: 'vulkan'
+      stable-svg:
+        toolchain: stable
+        features: 'svg'
+        exampleArgs: '--driver svg'
   variables:
     platform: ${{ parameters.platform }}
     image: ${{ parameters.image }}
@@ -61,7 +61,6 @@ jobs:
     displayName: Print versions
 
   # Note: features are ignored when set in the workspace. This is a known bug in cargo (#5015), so cd into skia-safe instead.
-  # Note: SKIA_BINARIES_KEY is set to empty ('') when the build script decides not to publish.
   - bash: |
       set -e
       cd skia-safe && cargo build --release --features "$(features)" --all-targets -vv
@@ -98,13 +97,12 @@ jobs:
     displayName: 'Release to GitHub rust-skia/skia-binaries'
     inputs:
       action: edit
-      gitHubConnection: pragmatrix
+      gitHubConnection: rust-skia-github-connection
       repositoryName: rust-skia/skia-binaries
       tagSource: manual
       target: 'master'
       tag: '$(SKIA_BINARIES_TAG)'
       assets: '$(Build.ArtifactStagingDirectory)/skia-binaries-$(SKIA_BINARIES_KEY).tar.gz'
       assetUploadMode: 'replace'
-      # TODO: change before merging this into master.
       isPreRelease: true
       addChangeLog: false
