+++ skia-bindings/build_support.rs
@@ -1,5 +1,6 @@
 //! Build support functions for the Rust-Skia library.
 
+pub mod azure;
 pub mod binaries;
 pub mod cargo;
 pub mod git;
